unified-medical-system/requirements.txt
Authlib
blinker
cachelib
certifi
cffi
charset-normalizer
click
cryptography
dnspython
Flask
Flask-Login
Flask-Mail
Flask-OAuthlib
Flask-PyMongo
Flask-Session
idna
itsdangerous
Jinja2
MarkupSafe
msgspec
oauthlib
pycparser
pymongo
python-dotenv
requests
requests-oauthlib
urllib3
Werkzeug
python-dateutil
python-barcode
pillow
transformers
reportlab
folium
gunicorn
qrcode
google-generativeai
aiohttp
Flask[async]